+++ packages/sdk/src/ConnectorClient.ts
@@ -3,6 +3,7 @@ import qs from "qs";
 import { ConnectorConfig } from "./ConnectorConfig";
 import {
     AccountEndpoint,
+    AttributesEndpoint,
     ChallengesEndpoint,
     FilesEndpoint,
     IncomingRequestsEndpoint,
@@ -16,6 +17,7 @@ import {
 
 export class ConnectorClient {
     public readonly account: AccountEndpoint;
+    public readonly attributes: AttributesEndpoint;
     public readonly challenges: ChallengesEndpoint;
     public readonly files: FilesEndpoint;
     public readonly incomingRequests: IncomingRequestsEndpoint;
@@ -44,6 +46,7 @@ export class ConnectorClient {
         });
 
         this.account = new AccountEndpoint(axiosInstance);
+        this.attributes = new AttributesEndpoint(axiosInstance);
         this.challenges = new ChallengesEndpoint(axiosInstance);
         this.files = new FilesEndpoint(axiosInstance);
         this.incomingRequests = new IncomingRequestsEndpoint(axiosInstance);