+++ packages/sdk/src/types/attributes/ConnectorAttributes.ts
@@ -1,0 +1,3 @@
+import { ConnectorAttribute } from "./ConnectorAttribute";
+
+export interface ConnectorAttributes extends Array<ConnectorAttribute> {}