+++ packages/sdk/src/types/attributes/index.ts
@@ -1,0 +1,5 @@
+export * from "./ConnectorAttribute";
+export * from "./ConnectorAttributes";
+export * from "./requests/CreateAttributeRequest";
+export * from "./requests/GetAttributesRequest";
+export * from "./requests/GetValidAttributesRequest";