+++ packages/sdk/src/types/attributes/requests/GetValidAttributesRequest.ts
@@ -1,0 +1,20 @@
+export interface GetValidAttributesRequest {
+    content?: {
+        "@type"?: string;
+        tags?: string[];
+        owner?: string;
+        key?: string;
+        isTechnical?: string;
+        confidentiality?: "public" | "private" | "protected";
+        value?: {
+            "@type"?: string;
+        };
+    };
+    succeeds?: string;
+    succeededBy?: string;
+    shareInfo?: {
+        requestReference?: string;
+        peer?: string;
+        sourceAttribute?: string;
+    };
+}